Extreme composition: a study
I was curious to try this after the second POODL course. I unwound the
inheritance hierarchy, pushing the conditionals back into the factory
method, and creating different types for each branch of the conditional.

Two patterns emerged: that the Container needs to switch on 1..n
bottles, and that the rest of them (description, action, next number)
all switch on 0..n bottles.

So I think we could have two entirely different BottleNumber classes for
the 0/n cases, composed of a container for the 1/n cases.